refactor(session): rename utils.ts to http-helpers.ts
