+++ apps/server/src/shared/http-helpers.ts
